Use IdentityCacheMapping to create cache key for pooling
Instead of using all the connection properties as the cache key, we will
use the well-established `IdentityCacheMapping`, which each JDBC
connector is supposed to provide.

The rationale is that the connection properties include also the
authentication credentials, which are not always guaranteed to be
static. For example, with AWS IAM authentication, the password is an
authentication token, which may change from one connection attempt to
another, which would force creation of a new connection pool each time,
defeating the purpose of connection pooling.

The reason for the existing implementation seems to be that the
different connection properties can change the behavior of the
connection, so there could be a case when we create a connection with
one set of properties, and later we reuse it even though the requested
set of properties is different and we return a connection that is
configured incorrectly for that. This does not seem to be a problem,
though, because each `CredentialPropertiesProvider` implementation
always creates the same set of properties.

The only exception to the above is the Snowflake connector, which
includes the warehouse name as part of its properties, and it should
also be part of the cache key. The solution in this case is to create a
custom implementation of `IdentityCacheMapping`, which wraps the
original `IdentityCacheKey` and adds the warehouse name to it.

Note that if user impersonation is disabled, the
`ExtraCredentialsBasedIdentityCacheMapping` implementation of
`IdentityCacheMapping` will be used, and it also includes the password
as part of the key; but this password is static, so is not a concern.

(cherry-picked from commit: 9bd56274077a3db61d686117d723561c5c68a0c6)

(cherry picked from commit 06e7187c0b05069e4fcadd110dd2b5c68fe20a8a)