Fix JNA dependency conflict in  Ranger plugin
There was a conflict on JNA dependency. It was dragged by
starburst-ranger-client and trino-main and these were using different
versions. Even that trino-main is separated with plugin by a separate
classloader and product was working just fine, we were not able to run
unit tests because there is no classloader separation.

To avoid the conflict the JNA dependency was removed from Ranger client
library and a dependency was added in the SEP in the same version that
is used in trino-main. Ranger client usage of JNA is very simple so
updating that library seems to be safe.

(cherry-picked from commit: 957a54fe79a49feda4d23c8457ed43efe054e58a)

(cherry picked from commit 853208d5fd967756f37a055006dafa597867c266)