Add Galaxy table level RBAC
Add AccessDenied explanations in GalaxyAccessControl

This commit adds Galaxy-specific explanations as extraInfo
in GalaxyAccessControl.  This required adding a couple of
AccessControlException overloadings for a few of the methods
that did not accept extraInfo.  Those changes have been submitted
as a Trino PR.

A typical message in Galaxy is:

Access denied: Cannot rename schema from catalog1.myschema1 to whackadoodle: Role public does not own the schema