Refactor 'experimental_parquet_optimized_writer_enabled' boolean to enum
The objectstore connector does not support connectors with the same
session property names but with different default values.
'experimental_parquet_optimized_writer_enabled' is false for hive but
true for delta.

Thus make the property an enum with a default value of 'DEFAULT'
that is interpreted differently by the hive/delta lake connectors.