Fix Trino UI display of queries
This commit adds a Role dropdown a the top of every
page in the Trino UI that is populated with the roles
the user could assume, with the user's current role
selected.  When the user changes the chosen role,
a DispatchResource request is made to get a new
dispatch token for the chosen role.

This is paired with a stargate commit that expects
the role in the /ui/api/query in DispatchResource.