Add random partition to buffer node mapping mode
Add random partition->buffer node mapping mode. It
should provide better performance for cases when Trino stage produces
a low number of partitions (corner case is arbitrary distribution when
just one partition is output). The previous implementation, when the
buffer node was pinned to the partition, would not work great for such a case.