Move encryption of pages to BufferExchangeSink
Previously encryption of pages to be sent to buffer service was
happening in SinkWriter. This resulted in poor throughput as SinkWriter
is single threaded. While having just a single thread is fine when it
comes to network bounded work, it does not work well with CPU bounded
work as encryption.