Do not allow reregistering buffer node with same id and different URI
This is safety measure in case new buffer node instance comes up and
tries to register itself with same ID. This is not bullet proof in any
way. It is still external responsibility to ensure that buffer nodes are
assigned globally unique ids. Yet this logic allows us to catch some set
of bugs.