Fix granting Galaxy view privileges in SQL
Before this commit, if you tried grant, deny or revoke
SELECT, UPDATE, INSERT or DELETE on a view using SQL in
Galaxy, the operation failed saying the "table" represented
by the view name did not exist.  After this commit, granting,
denying and revoking SELECT, UPDATE, INSERT or DELETE on
a view works correctly in Galaxy.

This commit includes a test that shows that granting,
denying and revoking works correctly.,