Do not enforce offline Maven for Galaxy
Using maven in offline mode does not always work. It happens that some
dependencies (which are not explicitly specified as <dependency> in
pom.xmls) are not downloaded by `mvn go-offline` and later on using
mvn --offline results in build failures.

Instead fighting which individual problem we decided to not be strict
about not allowing maven to download artifacts during build.