Reserve memory for chunk data before read request
Account for memory which will be used by the data pages in chunk
yet to be read before we issue read request to data-server.
That ensures that we will not overshoot the memory limit for
BufferExchangeSource which currently is possible when we issue multiple
concurrent read requests and each of those reads a lot of data.