Support row-level filters in Galaxy
This PR gets row-level filters from the EntityPrivileges
object returned when galaxy-trino requests table privileges.
The identity used to get table privileges for the row
filter when the row filter expression is analyzed is the role
that owns the row filter.  That roleId is encoded in the
user string of the ViewExpression object created when
the row filter is analyzed.

This commit initializes the AesCrypto at the time of
the first call to avoid a JVM crash when it is initialized
statically.