Guess table type based on observed tables in OSC
`ObjectStoreConnector`'s `ObjectStoreMetadata.getTableHandle` is
expensive when table type is not known upfront, as it invokes underlying
connectors, effectively reading the metastore multiple times.

This commit improves the behavior for previously-unseen tables by
guessing their type based on tables seen previously. This should improve
user experience for users that have existing data sets of e.g. Hive or
Delta tables, by improving the behavior when exact information is not
present in the `TableTypeCache`. This comes at the cost of (previously
favored) Iceberg tables.