Add error handling for saturated floor cast
An exception wasn't thrown until today because of rule ordering.
For example, double NaN can't be cast to smallint. Given the query
"SELECT * FROM (VALUES SMALLINT '1') t(a) WHERE a IS DISTINCT FROM nan()",
the filter's predicate is replaced with TRUE by
UnwrapCastInComparison.FilterExpressionRewrite before hitting
DomainTranslator. When manually disabling this rule, a TrinoException
with the message "Unable to cast double NaN to smallint"
is thrown by DoubleOperators#saturatedFloorCastToLong.