Support column masks in Galaxy
This commit adds support for Galaxy column masks, and
a mimimal test that demonstrates that masking works,
and that the role owning the mask determines privileges
when the column mask expression is analyzed.

This commit also makes the changes required because
EntityPrivileges got new member columnMasks, and because
TestingAcocuntClient tagging API changed to accomodate
tagging any contained entity in galaxy-trino tests.

Add column masking functions, lots of small fixes