Cache table type in ObjectStore connector
The `TableTypeCache` caches `TableType` based on table names, so that we
known which metadata to consult. In case the cached value is incorrect,
we continue asking other metadatas. The assumption is that order of
metadata queries matters for performance, but doesn't affect the end
result. If the metadatas didn't cache, it would obviously be true, but
the assumption is that the cache eviction added in
a963a24a22ee8cf125b056af0053b2f32eb83445 guarantees consistency.

The additional value of `TableTypeCache` is that it detects "catalog
affinity". Conceptually, if all or almost all tables are Delta tables,
we can assume a previously unseen table to be probably a Delta table
too.

This mostly reverts 7952669dd936f2096c8cf7bfe33a86ad576ecbbd and
9358dd33b870839e4b93e23ffd487cb8146f59eb, reintroducing
df02727b535dac0cf584722110aa5ee492a6478a and
2c60d48d3d057b7b162515a8fbdcb92c558ca741.

Co-authored-by: Konrad Dziedzic <konraddziedzic@gmail.com>