Refactor session property init in ObjectStoreConnector
Move it out from the constructor to a helper method.
Doing it eagerly (before the constructor returns) lets this be
test-covered by `TestObjectStorePlugin`. Doing it in a helper method is
necessary for readability if other properties eagerly-initialized.

Helper class `DelegateConnectors` is introduced to avoid passing 4x
Connector as positional arguments.