Eliminate round trips needed to run a Galaxy query
Comment out the call in AccessControlManger.filterColumns
to AccessControlManager.filterTables.  The call is supposed to
be an optimization, but it requires two additional round trips
from Trino to the portal-server, and we get complete privileges
information from the call to get EntityPrivileges in
GalaxyAccessControl.filterColumns.

Obviously commenting out three lines of AccessControlManager
might not be the most tasteful way to make the change.
Suggestions welcome.