Introduce connector cache for Metadata Service
Usage shows that clients will interact with the Metadata Service
for the same catalog over the course of a prolonged period. In
these instances re-using the connector will save in time to
setup as well as memory (currently the metadata service leaks
memory on every usage). The cache duration is short so that
one time uses will not retain memory.

Only the NORMAL connector is retained as the system and information
schema connectors do not involve real plugins.