Add support for multiple issuers in GalaxyAuthenticatorController
Separate out GalaxyAuthenticatorController common code into one helper
class for static functions and an abstract implementation that moves
issuer and audience checks from the JwtParser to the abstract common
controller code. This will allow the GalaxyMetadataAuthenticator to
create a specialization of the abstract class to check for multiple
issuers in a future commit. Further, any single issuer will be able to
bind zero or more audiences. In the future, at least one audience will
be required by the GalaxyAuthenticatorController after requests and
Controller configs are updated.

GalaxyAuthenticatorController will continue to have a single issuer but
in a subsequent commit, the GalaxyMetadataAuthenticatorController will\
support 2 issuers (portal and dispatcher).

Moved common code in TestGalaxyAuthenticator to static functions.