Eliminate calls to listEnabledRoles, part deux
This commit eliminates calls to listEnabledRoles
when creating a Trino session if the access control
is GalaxyAccessControl and the operation is a DML
operation.  An earlier attempt failed, because it was
assuming that the AccessControl instance in
QuerySessionSupplier is the SystemAccessControl
instance but it is not.  This commit gets the list of
SystemAccessControl instances from AccessControlManager,
and omits the call to listEnabledROles only if the operation
is a DML operation, and if GalaxyAccessControl is
the only instance in the list.