Break reference chain to input-stream ReadListener.
This commit breaks reference chain from Jetty's (implementation of
ServletInputStream) to registered ReadListener.

For some reason Jetty keeps reference to ReadListener attached to
ServletInputStream even after releases is already
complete. We need to break reference chain as ReadListener we use
has references to Slice used for holding request data
while at this point this memory is no longer accounted for in
MemoryAllocator. This was resulting in OOMs.