Add PlanSignature#dynamicPredicate
Dynamic predicate is predicate dervied from  dynamic filter that was applied on
top of enforced PlanSignature#predicate. Dynamic predicate is not enforced, but best effort.
You can use output of `PlanSignature A` to derive output of matching `PlanSignature B`
as long as `B.dynamicPredicate` is a subset of `A.dynamicPredicate`.