Validate iceberg storage_schema MV privileges in Galaxy
This commit validates the optional storage_schema property when creating materialized views with Iceberg.

Galaxy requires CREATE_TABLE privileges on a schema when creating a materialized view. This commit checks for the CREATE_TABLE privileges on the storage_schema when it is set to prevent a materialized view from being created in a schema the role isn't allowed to create a table in.