Revert "Enable connection pooling in Synapse test"
This reverts commit a3d1b5daac4ecd9bfa1713a1240683f7dc485c26.

After this change Synapse tests very frequently run into "Connection
reset" errors.

If we hold a connection for a long duration and aren't actively
executing a query or a transaction on the connection Synapse can
terminate the connection forcefully via a TCP RST packet. This manifests
as a "Connection reset" error on the JDBC driver.

Hikari doesn't actively track health of connections which have been
taken from the pool, only the connections within the pool are tracked.
This means that when the next operation is attempted on such a
connection the error bubbles up to the application and since we don't
have any retries in JDBC page sink or record cursor such errors lead to
query failures.

Note that this suggests that all JDBC connectors can be similarly
affected (and this was verified with Postgres connector and a low
idle timeout on Postgres side) and the proper fix would be to add
retries on transient errors in the JDBC page sink and record cursor.

(cherry-picked from commit: e51ad0f0281e7d2cd7457e1abc4dfc20f5c81961)

(cherry picked from commit 4ddb33859be5332020487c6e7ffadb4e444330cf)