Introduce Galaxy Caching for information_schema
Introduces an optional Galaxy caching system that redirects
some information_schema queries to Galaxy so that it can return
the results from its cache. This takes advantage of the introduced
injection of InformationSchemaPageSourceProvider.

Additionally replaces `TableCommentSystemTable` with
`GalaxyCacheTableCommentSystemTable` so that queries against
system.metadata.table_comments can be redirected to the cache.