Do more concurrent work in parallel streamTableColumns
Getting initial iterator is cheap operation (and can actually be a
no-op, depending on the implementation). Iterating it is what is
expensive. The V1 mode was supposed to be parallel, so make it more
parallel.

Not introducing new mode, changing V1 directly, because V1 didn't
deliver on its improvement promise and would otherwise be scrapped.
"V1"  was not supposed to mean a version set in stone. We would call it
"parallel" but we didn't want to expose implementation hints to
customers, chose "V1" as an opaque name.