Cache catalog visibility within a query
This commit adds ContentVisibility catalogVisibility in the
GalaxyQueryPermissions class.  That class stores a variety of per-query
and per-session information.  This commit changes
GalaxySystemAccessController.getCatalogVisibility to get the visiblity
from the GalaxyPermissionsCache.

This guarantees that for each query and each roleId assumed in the query
that catalog visibility will be fetched at most once.