Fix forwarding of ForwardingSystemAccessControl.filterViewQuery()
Make `ForwardingSystemAccessControl` truly forwarding again in the fork.
The method that lacked forwarding is Galaxy fork differentiator
(attributed to 1749c197a953c7c758301ca28e7196bcce0edb1b commit). This
hides default method defined in `SystemAccessControl` and relies on
whatever delegate provides.

The `ForwardingSystemAccessControl` is not used in production code
(except for `FileBasedSystemAccessControl` with file refresh, which is
not used in Galaxy produciton), so the change should have no effect
outside of tests.