Change Hive CSV escape character to backslash when set to double quote
Hive has a bug that changes the escape character to backslash when it is
explicitly set to a double quote during deserialization.  The native CSV
reader is a bug for bug compatible reimplementation of the Hive OpenCSV
reader, so this behavior must be replicated.

(cherry picked from commit 006097fe4cba4c866aaae98efa1466a2cb0b302e)