Allow re-using PostgreSQL container in tests
Ability to iterate faster on integration test development for PostgreSQL by adding the support for TESTCONTAINERS_REUSE_ENABLE environment variable.

This is based on previous commit fbd1de0 by @kokosing, which adds the feature and enables it for SQL Server and MySQL. Based on further comments from @kokosing, enablement of this feature for specific database tests needs to be tested one-by-one as not all test containers seem to digest that change without problems. This justifies adding it iteratively for selected databases based on a need-to-use basis.

(cherry picked from commit 64000d2aa0d5d364f40ad4a89fe9965e40411339)