Skip getCatalogVisibility when proof already known
`catalogVisibility` endpoint returns a list of catalogs where there
exists, or can exist, a relation, to which user has some privileges,
either explicit or implicit, and not overridden with DENY.

In some cases, the call is redundant, if, withing context of particular
query and identity, we have already seen a table in given catalog.

This is especially useful for `DESCRIBE`, where `checkCanShowColumns` is
invoked first, giving the proof of catalog visibility.