Update memory requirements for pending and running tasks
If we observe new memory requirements for a stage we update the value
for tasks which are running and waiting for node lease. That ensure that
state of NodeAllocator better reflects actual usage of memory on the
cluster.

(cherry picked from commit 2e6ede0665c7d12ba76fddfe007e6aa1b1016585)