Retrieve table comments in JDBC connectors in bulk
This basically retrofits exists `BaseJdbcClient.getTableNames` into "get
table names and comments", because retrieving comments does not require
any additional calls.

Of course, the small conceptual change is adorned with lots of interface
changes, new DefaultJdbcMetadata code, session toggles and such, which
make it far more baroque in perception.

The functionality is enabled by default, with a config and session
properties acting as a kill switch. Besides the default DMA mode, there
is also DMA_P mode which is probably not going to be useful. It's there
for experimentation.