Fix Delta memory leak via ObjectStore
Pass `beginQuery` and `cleanupQuery` calls to all delegate connectors.
Previously this was called for Hive only as historically only Hive
needed that. However there was no automated way to detect other delegate
connectors do not gain the method. In fact, Delta recently did and lack
of these invocations means a serious memory leak in the connector.