Add getNextFilteredRowRanges SPI to ConnectorPageSource
Returns a list of row number ranges that will be produced by this ConnectorPageSource.
For example, suppose the row numbers returned for a split are `[0, 1, 2, 4, 5, 7, 8, 9]`,
then this method should return 3 row ranges: `[0-2], [4-5], [7-9]`.
This method is intended to be used by warp speed connector to improve queries
when some columns are in the cache while others need to be fetched from source.

(cherry picked from commit 6cbf60f66143329bee09eae934d501e6086d381d)