Galaxy Catalog Connector
Specialized connector intended to proxy/contain specialized needs
or behavior for Galaxy.

The first feature included is support for SQL Language Functions.
Implements the new ConnectorMetadata methods for `listLanguageFunctions`
etc. The implementation in this connector is merely a proxy that
makes a REST call using a client provided by Galaxy.