Add scheduler span under query that aggerates all stages
The Trino spans below the the Query span are Dispatch, Analyzer, Planner
and then a sequence of Stage spans. Placing the set of Stage spans under
a new Scheduler span allows all execution spans to be grouped together.
In addition, new Exchange spans are also being added to instrument the
exchanges that operate across stages.

Normalize getter in ExchangeContext interface.