Pass failoverHappened to ConnectorPageSourceProvider
No clients are using this info for now, so add a new method that
defaults to ignoring the extra parameter.  This minimizes changes, as we
don't need to extend the two dozen interface implementations.  But in
the future, we can simplify this to just one method and let each
implementation decide whether to use splitAddressEnforced or ignore it.

(cherry picked from commit 2fae4a719ffea5348e44421bdaeaf84131fd3ebd)