Enable Objectstore connector to work with Trino native file systems
Add missing bindings to the Objectstore connector to enable
it to make use of the native TrinoFileSystem implementations.

Use the native s3 TrinoFileSystem in testing the Objectstore plugin,
but keep using the hdfs based TrinoFileSystem in the bulk of the
tests of the connector (the test classes based on
BaseObjectStoreConnectorTest) to match the reality from production.