Remove duplicated TestDynamoDbConnectorTest execution
Both `starburst-dynamodb` and `trino-galaxy-dynamodb` were running this
test. This was a left over from times where the main connector
implementation lived in a separate repository, the shared plugins repo.
Additional execution of the test within Galaxy was different than the
one from the shared plugins repo. The galaxy execution had Galaxy e.g.
forked main, SPI, base JDBC modules. But that's thing of the past.

Now that the base connector lives in this repo, the execution differs
only in the classpath, but there is no classpath scanning, so that
cannot influence the result.