Remove duplicated TestSynapseConnectorTest execution
Both `starburst-synapse` and `trino-galaxy-synapse` were running
`TestSynapseConnectorTest`. This was a left over from times where
Synapse lived in a separate repository. Additional execution of
`TestSynapseConnectorTest` within Galaxy was different than the one from
the shared plugins repo. The galaxy execution had Galaxy forked base
JDBC and SQL Server connectors. But that's thing of the past.

Now that Synapse lives in this repo, the execution differ only in the
classpath, but there is no classpath scanning, so that cannot influence
the result.

This also fixes test execution of other tests that could exist in
`trino-galaxy-synapse` module. However, none exist today.