Remove duplicated SQL Server tests execution
Both `starburst-sqlserver` and `trino-galaxy-sqlserver` were running
these tests. This was a left over from times where the main connector
implementation lived in a separate repository, the shared plugins repo.
Additional execution of the tests within Galaxy was different than the
one from the shared plugins repo. The galaxy execution had Galaxy e.g.
forked main, SPI, base JDBC modules. But that's thing of the past.

Now that the base connector lives in this repo, the execution differs
only in the classpath, but there is no classpath scanning, so that
cannot influence the result.

This also fixes test execution of tests other than
`TestStarburstSqlServerPlugin` in `trino-galaxy-sqlserver`, which would
not be attempted on CI (currently none exists).