Remove Galaxy Cache
It was quickly built for Yello but they didn't end up using it. There's
no reason to maintain it. If we want something similar there are better
ways to do it.

This will require removing a Feature Flag on the Galaxy side. No
customer is using it right now but the Feature Flag should be globally
disabled first just in case.