Avoid pre-projecting literals in window measures
The frame end expression was being pre-projected so that
it could be rewritten into a symbol reference, as that's
the only type of expression WindowNode supports.

When the frame end is a constant and a measure expression
contains the same constant, it gets extracted and rewritten
as a symbol reference withing the measure expression. This
is unnecessary at best and a performance issue in the worst
case.

(cherry picked from commit a2748b3fa26563ef713cde450b32f6081d29ae3b)
(cherry picked from commit 808d4450897090d75fdcbc9e83ad6dd8d0d07e38)