Fix potentially incorrect table read version for writes on Delta Lake
In the context of concurrent operations, it may potentially happen
that between the time when the version of table has been initially
read for scanning and the time when the planning of the write
operation begins that a new version of the table is current.
Nevertheless, use in performing the write operations the version
of the table initially read as reference for beginning the
operation to ensure the consistency of the operation.

This strategy helps in unintentionally skipping the processing
of transaction log information corresponding to other concurrent
operations.

(cherry picked from commit 11eedf4a36fc0a61b4191e4946b5c8f21d8ed460)